+++ BetterCrafting/ModConfig.cs
@@ -4,6 +4,8 @@ using StardewModdingAPI.Utilities;
 
 using Leclair.Stardew.BetterCrafting.Models;
 
+using Leclair.Stardew.Common.Enums;
+using Leclair.Stardew.Common.Inventory;
 using Leclair.Stardew.Common.Types;
 
 namespace Leclair.Stardew.BetterCrafting;
@@ -48,7 +50,7 @@ public class ModConfig {
 	public TTWhen ShowKeybindTooltip { get; set; } = TTWhen.ForController;
 
 	public KeybindList SuppressBC { get; set; } = KeybindList.Parse("LeftShift");
-
+	public KeybindList ModiferKey { get; set; } = KeybindList.Parse("LeftShift, RightTrigger");
 	public KeybindList FavoriteKey { get; set; } = KeybindList.Parse("F, ControllerBack");
 	public KeybindList BulkCraftKey { get; set; } = KeybindList.Parse("None");
 	public KeybindList SearchKey { get; set; } = KeybindList.Parse("F3");
@@ -71,12 +73,27 @@ public class ModConfig {
 
 	// Better Workbench
 	public bool UseDiscovery { get; set; } = true;
-	public int MaxInventories { get; set; } = 15;
+	public int MaxInventories { get; set; } = 32;
 	public int MaxDistance { get; set; } = 20;
 	public int MaxCheckedTiles { get; set; } = 500;
 
 	public bool UseDiagonalConnections { get; set; } = true;
 
 	public CaseInsensitiveHashSet ValidConnectors { get; set; } = new();
+
+	// Transfer Behavior
+
+	public bool UseTransfer { get; set; } = true;
+	public Behaviors AddToBehaviors { get; set; } = new Behaviors();
+	public Behaviors FillFromBehaviors { get; set; } = new Behaviors();
+}
+
+public class Behaviors {
+
+	public TransferBehavior UseTool { get; set; } = new TransferBehavior(TransferMode.All, 1);
+	public TransferBehavior UseToolModified { get; set; } = new TransferBehavior(TransferMode.AllButQuantity, 1);
+
+	public TransferBehavior Action { get; set; } = new TransferBehavior(TransferMode.Quantity, 1);
+	public TransferBehavior ActionModified { get; set; } = new TransferBehavior(TransferMode.Half, 1);
 
 }